requirements.txt
# Core AI/LLM Libraries
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.20

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0

# Web and API
requests>=2.31.0
python-dotenv>=1.0.0

# CLI and Interface
rich>=13.0.0
click>=8.1.0
prompt-toolkit>=3.0.0

# Utilities
python-dateutil>=2.8.0
jsonschema>=4.17.0
pyyaml>=6.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
